+++ src/mapper.js
@@ -1,24 +1,20 @@
 import { buildCallMethod } from './internals';
-import { createPromiseBrowserHttpLayer, createPromiseResponseHandler } from './http-layers/browser-promise-layer';
+import { createDefaultHttpLayer } from './http-layers/default';
 
 /**
  * Creates an api mapper object
  * @param {Object} store - The redux store object
  * @param {Object} config - An object defining api mapper
  * @param {Object} [httpLayer] - An object for resolving http requests
- * @param {Function} [httpResponseHandler] - A function to handle the response from the http layer and dispatch new request state
  */
-export function createMapper(store, config, httpLayer, httpResponseHandler) {
+export function createMapper(store, config, httpLayer) {
   let mapper = {};
 
-  httpLayer           = httpLayer || createPromiseBrowserHttpLayer();
-  httpResponseHandler = httpResponseHandler || createPromiseResponseHandler();
-
-  mapper.host                = config.host;
-  mapper.headers             = config.headers;
-  mapper.httpResponseHandler = httpResponseHandler;
-  mapper.httpLayer           = httpLayer;
-  mapper.store               = store;
+  mapper.httpLayer = httpLayer || createDefaultHttpLayer();
+  mapper.host      = config.host;
+  mapper.headers   = config.headers;
+  mapper.options   = config.options;
+  mapper.store     = store;
 
   config.resources.forEach(
     resource => addResourceToMapper(mapper, resource)
@@ -33,7 +29,7 @@ export function createMapper(store, config, httpLayer, httpResponseHandler) {
  * @param {Object} resource - An object resource
  */
 export function addResourceToMapper(mapper, resource) {
-  mapper[resource.name] = {path : resource.path, headers : resource.headers};
+  mapper[resource.name] = {path : resource.path, headers : resource.headers, options : resource.options};
   resource.endPoints.forEach(
     endPoint => addEndPointToMapper(mapper, resource, endPoint)
   );
@@ -51,6 +47,7 @@ export function addEndPointToMapper(mapper, resource, endPoint) {
     headers : endPoint.headers,
     method  : endPoint.method,
     action  : endPoint.action,
+    options : endPoint.options,
     call    : buildCallMethod(mapper, resource, endPoint)
   };
 }